[Refactor] 確認ダイアログを AppConfirmDialog に共通化
- AppConfirmDialog を新規作成（title?・message・confirmLabel・isDestructive・confirmColor）
- spot_detail_page の _SaveConfirmDialog / _DeleteConfirmDialog / _MoveToWentConfirmDialog を削除
- walk_route_page の _RouteDeleteConfirmDialog を削除
- settings_page の _ConfirmActionDialog を削除
- 各ページから AppConfirmDialog を使用するよう変更
